requirements.txt
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
sqlalchemy
pydantic
fastapi
psycopg2-binary
uvicorn
